agenticpdf/requirements.txt
python-dotenv
llama_index
pytest==8.3.4
pytest-mock==3.6.1
openai
pytest
pytest-asyncio
Flask

llama-index-core
llama-index-utils-workflow
llama-index-llms-openai
llama-parse
llama-index-embeddings-openai
llama-index-readers-whisper

gradio
